JAV-508 & SCB-11 move metric dto from metrics-core to metrics-common and add metrics publish via rest schema
