+++ metrics/pom.xml
@@ -34,5 +34,6 @@
   <modules>
     <module>metrics-core</module>
     <module>metrics-sample</module>
+    <module>metrics-common</module>
   </modules>
 </project>
